serenity_project/requirements.txt
asgiref==3.5.2
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
coverage==6.5.0
coveralls==3.3.1
Django==3.2.16
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
docopt==0.6.2
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
importlib-metadata==3.7.0
install==1.3.5
mccabe==0.7.0
mypy-extensions==0.4.3
numpy==1.21.6
pandas==1.3.5
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.4
pycodestyle==2.9.1
pyflakes==2.5.0
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
style==1.1.0
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
update==0.0.1
urllib3==1.26.12
whitenoise==6.2.0
zipp==3.9.0